+++ src/main/java/org/logstash/tcp/SslContextBuilder.java
@@ -18,22 +18,33 @@ import org.bouncycastle.pkcs.PKCS8EncryptedPrivateKeyInfo;
 
 import javax.crypto.Cipher;
 import javax.net.ssl.SSLException;
+import javax.net.ssl.SSLServerSocketFactory;
 import java.io.FileInputStream;
 import java.io.FileReader;
 import java.io.IOException;
 import java.security.PrivateKey;
+import java.security.Security;
 import java.security.cert.Certificate;
 import java.security.cert.CertificateException;
 import java.security.cert.CertificateFactory;
 import java.security.cert.X509Certificate;
 import java.util.ArrayList;
+import java.util.Arrays;
+import java.util.HashSet;
 import java.util.List;
+import java.util.Set;
 
 public class SslContextBuilder {
 
     private static final String[] NULL_STRING_ARRAY = new String[0];
 
     private final static Logger logger = LogManager.getLogger(SslContextBuilder.class);
+
+    public static List<String> getSupportedCipherSuites() {
+        return Arrays.asList(
+            ((javax.net.ssl.SSLServerSocketFactory) SSLServerSocketFactory.getDefault()).getSupportedCipherSuites()
+        );
+    }
 
     private boolean sslEnabled;
     private boolean shouldVerify;
@@ -45,6 +56,9 @@ public class SslContextBuilder {
 
     private String[] certificateAuthorities = NULL_STRING_ARRAY;
     private String[] extraChainCerts = NULL_STRING_ARRAY;
+
+    private String[] supportedProtocols = NULL_STRING_ARRAY;
+    private String[] cipherSuites = NULL_STRING_ARRAY;
 
     public SslContextBuilder setSslEnabled(boolean enabled) {
         this.sslEnabled = enabled;
@@ -81,6 +95,27 @@ public class SslContextBuilder {
         return this;
     }
 
+    public SslContextBuilder setSslSupportedProtocols(String[] protocols) {
+        this.supportedProtocols = protocols;
+        return this;
+    }
+
+    public SslContextBuilder setSslCipherSuites(String[] suites) {
+        if (suites.length > 0) {
+            final Set<String> supportedCipherSuites = new HashSet<>(getSupportedCipherSuites());
+            for (String cipher : suites) {
+                if (supportedCipherSuites.contains(cipher)) {
+                    logger.debug("{} cipher is supported", cipher);
+                } else {
+                    throw new IllegalArgumentException("Cipher `" + cipher + "` is not available");
+                }
+            }
+        }
+
+        this.cipherSuites = suites;
+        return this;
+    }
+
     public SslContext buildContext() throws Exception {
         if (!sslEnabled) return null;
 
@@ -90,6 +125,9 @@ public class SslContextBuilder {
         if (keyPath == null) {
             throw new IllegalArgumentException("missing ssl_key");
         }
+
+        // NOTE: decrypting openssl key-pair (PEMEncryptedKeyPair) assumes the BC provider
+        installBouncyCastleProvider();
 
         // Check key strength
         if (Cipher.getMaxAllowedKeyLength("AES") <= 128) {
@@ -148,6 +186,9 @@ public class SslContextBuilder {
 
         sslContextBuilder.clientAuth(shouldVerify ? ClientAuth.REQUIRE : ClientAuth.NONE);
 
+        if (supportedProtocols.length > 0) sslContextBuilder.protocols(supportedProtocols);
+        if (cipherSuites.length > 0) sslContextBuilder.ciphers(Arrays.asList(cipherSuites));
+
         try {
             return sslContextBuilder.build();
         } catch (SSLException e) {
@@ -187,6 +228,14 @@ public class SslContextBuilder {
             fis.close();
         }
         return certificates;
+    }
+
+    private static void installBouncyCastleProvider() {
+        synchronized (Security.class) {
+            if (Security.getProvider(BouncyCastleProvider.PROVIDER_NAME) == null) {
+                Security.addProvider(new BouncyCastleProvider());
+            }
+        }
     }
 
 }